villeController, CommandeController

+++ back-end/database/migrations/2024_04_02_195024_create_factures_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('facture', function (Blueprint $table) {
-            $table->unsignedBigInteger('idFac');
+            $table->unsignedInteger('idFac');
             $table->primary('idFac');
             $table->datetime('dateFac');
             $table->decimal('remise', 8, 2);
